Send iMessage and SMS from the terminal command line
File used to send quick sms or iMessage from the terminal command line.
If iMessage service is not available it uses SMS. If that’s not
available it returns with failure. Can execute script either through
command line or in script editor - or can be compiled into an app - is
not suitable for json or python/perl/php calls.